backend/requirements.txt
asgiref==3.6.0
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
install==1.3.5
markdown2==2.4.8
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.0
tzdata==2023.3
whitenoise==6.4.0
